# backend/02-presigned-url-upload/chatgpt/app/openapi.yaml
openapi: 3.1.3
info:
  title: Posts Presigned API
  version: "1.0"
servers:
  - url: http://localhost:8080/
paths:
  /api/uploads/init:
    post:
      summary: Get presigned upload info (POST or PUT)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - filename
              properties:
                filename:
                  type: string
                  description: Name of the file to upload
                method:
                  type: string
                  enum: [post, put]
                  description: Upload method (POST or PUT)
      responses:
        "200":
          description: Presigned upload info returned
          content:
            application/json:
              schema:
                type: object
                properties:
                  objectKey:
                    type: string
                    description: Key of the object to be uploaded
                  url:
                    type: string
                    description: Presigned URL for PUT or POST
                  formData:
                    type: object
                    additionalProperties: true
                    description: Required form fields for POST uploads

  /api/posts:
    post:
      summary: Create a post referencing an already-uploaded object
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
                - object_key
              properties:
                title:
                  type: string
                description:
                  type: string
                object_key:
                  type: string
                  description: Key of the uploaded object in MinIO
      responses:
        "201":
          description: Post created successfully

    get:
      summary: List all posts
      responses:
        "200":
          description: List of posts
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    title:
                      type: string
                    description:
                      type: string
                    object_key:
                      type: string

  /api/posts/{id}:
    get:
      summary: Get a post by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Post details
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  title:
                    type: string
                  description:
                    type: string
                  object_key:
                    type: string

    delete:
      summary: Delete a post by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Post deleted successfully

components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-KEY

security:
  - ApiKeyAuth: []